feat: Implement admin dashboard layout and routes; add protected admin routes for inquiries, programs, news, sections, and events management
